update to settings page
